pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chimera-spawn"
version = "1.0.0"
description = "Modern systemd-nspawn container orchestration system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Chimera Development Team", email = "dev@chimera.local"},
]
keywords = ["systemd", "nspawn", "containers", "orchestration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]

# Ubuntu 24.04 compatible versions
dependencies = [
    "pydantic>=1.10.14,<2.0",
    "typer>=0.9.0",
    "rich>=13.7.1",
    "dbus-next>=0.2.3",
    "python-systemd>=235",
    "ruamel.yaml>=0.17.21",
    "jinja2>=3.1.2",
    "watchfiles>=0.21.0",
    "aiofiles>=23.2.1",
    "httpx>=0.26.0",
    "psutil>=5.9.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
chimeractl = "chimera.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
chimera = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
